add idles and display_string

+++ bonus/sources_bonus/parsing/get_map.c
@@ -6,7 +6,7 @@
 /*   By: twang <twang@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/08 19:13:46 by wangthea          #+#    #+#             */
-/*   Updated: 2023/03/10 12:38:51 by twang            ###   ########.fr       */
+/*   Updated: 2023/03/13 17:24:47 by twang            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,9 +22,9 @@ static void	read_error(int fd)
 static int	get_allocation_size(t_game *g, const char *file)
 {
 	int	size;
-	int		fd;
-	int		nb_bytes_read;
-	int8_t		c;
+	int	fd;
+	int	nb_bytes_read;
+	int	c;
 
 	fd = open_file(file);
 	nb_bytes_read = 1;
@@ -45,11 +45,11 @@ static int	get_allocation_size(t_game *g, const char *file)
 
 static char	*fill_map(t_game *g, const char *file)
 {
-	int	i;
+	int		i;
 	int		fd;
 	int		nb_bytes_read;
-	int8_t		c;
-	char		*str;
+	int		c;
+	char	*str;
 
 	str = ft_calloc(get_allocation_size(g, file) + 1, sizeof(char));
 	if (!str)
